refactor(web): split tool-strategy into per-agent files (2/2)
Move cursor/pi/zcode builders into tool-strategy/, relocate the
agent-agnostic normalize functions to tool-normalize.ts and
formatMessageTime to lib/format.ts, then delete the legacy
tool-strategy.ts shell. Consumer imports of ./tool-strategy now
resolve to the directory index.

Issue: CS-19 (step 2/2)
