+++ apps/web/src/lib/format.ts
@@ -36,3 +36,31 @@ export function formatTokens(n: number) {
   if (n >= 1_000) return `${(n / 1_000).toFixed(1)}K`;
   return n.toString();
 }
+
+export function formatMessageTime(rawTime: number | string) {
+  if (typeof rawTime === "number" && rawTime <= 0) return "Unknown time";
+
+  let date: Date | null = null;
+  if (typeof rawTime === "number") {
+    const normalized = rawTime < 10 ** 12 ? rawTime * 1000 : rawTime;
+    date = new Date(normalized);
+  } else if (typeof rawTime === "string") {
+    if (rawTime.trim()) {
+      const timestamp = Number(rawTime);
+      if (!Number.isNaN(timestamp) && timestamp > 0) {
+        date = new Date(timestamp < 10 ** 12 ? timestamp * 1000 : timestamp);
+      } else {
+        date = new Date(rawTime);
+      }
+    }
+  }
+
+  if (!date || Number.isNaN(date.getTime())) return "Unknown time";
+
+  return date.toLocaleTimeString("en-US", {
+    hour: "2-digit",
+    minute: "2-digit",
+    second: "2-digit",
+    hour12: true,
+  });
+}
